requirements.txt
fastapi==0.115.8
uvicorn==0.42.0
langgraph==1.1.2
langchain==1.2.12
pymilvus==2.6.1
transformers==4.44.2
pillow==12.1.1
numpy==2.4.3
requests==2.32.5
python-multipart==0.0.22
aiofiles==25.1.0
aiohttp
langsmith==0.7.17
openai==2.28.0
langchain-openai==1.1.11
duckduckgo-search==8.1.1
pydantic==2.12.5
python-dotenv==1.2.2
modelscope==1.35.0
SQLAlchemy==2.0.30
pymysql==1.1.0